requirements.txt
aws-shell==0.2.2
awscli==1.29.62
boto3==1.28.62
botocore==1.31.62
colorama==0.4.4
configobj==5.0.8
docutils==0.16
jmespath==1.0.1
prompt-toolkit==1.0.18
pyasn1==0.5.0
Pygments==2.16.1
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.6
wcwidth==0.2.8
aws-shell==0.2.2
awscli==1.29.62
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
click==8.1.7
colorama==0.4.4
configobj==5.0.8
docutils==0.16
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==10.0.1
prompt-toolkit==1.0.18
pyasn1==0.5.0
Pygments==2.16.1
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
aws-shell==0.2.2
awscli==1.29.62
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
click==8.1.7
colorama==0.4.4
configobj==5.0.8
docutils==0.16
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==10.0.1
prompt-toolkit==1.0.18
pyasn1==0.5.0
Pygments==2.16.1
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
aws-shell==0.2.2
awscli==1.29.62
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
click==8.1.7
colorama==0.4.4
configobj==5.0.8
docutils==0.16
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==10.0.1
prompt-toolkit==1.0.18
psycopg2-binary==2.9.9
pyasn1==0.5.0
Pygments==2.16.1
python-dateutil==2.8.2
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
